34: Move error types
