+++ rag-toolchain/src/clients/mod.rs
@@ -3,3 +3,5 @@ mod model;
 pub mod openai_client;
 /// Module that defines the traits used by the clients in this crate
 pub mod traits;
+
+pub mod types;
