+++ rag-toolchain/src/clients/model/mod.rs
@@ -1,2 +1,3 @@
 pub mod chat_completions;
 pub mod embeddings;
+pub mod errors;
