+++ rag-toolchain/src/clients/traits.rs
@@ -2,6 +2,8 @@ use crate::common::types::{Chunk, Chunks, Embedding};
 use async_trait::async_trait;
 use std::error::Error;
 
+use super::types::PromptMessage;
+
 /// # AsyncEmbeddingClient
 /// Trait for any client that generates embeddings asynchronously
 #[async_trait]
@@ -12,4 +14,15 @@ pub trait AsyncEmbeddingClient {
         &self,
         text: Chunks,
     ) -> Result<Vec<(Chunk, Embedding)>, Self::ErrorType>;
+}
+
+/// # AsyncEmbeddingClient
+/// Trait for any client that generates chat completions asynchronously
+#[async_trait]
+pub trait AsyncChatClient {
+    type ErrorType: Error;
+    async fn invoke(
+        &self,
+        prompt_messages: Vec<PromptMessage>,
+    ) -> Result<PromptMessage, Self::ErrorType>;
 }
